Add GITHUB_TOKEN envvar in metadata-action steps
It seems the metadata-action step requires a Github token to work.
I'm guessing it has to talk to the Github API to get some release info.

It only needs to be added here to make it work with `act` for local
testing, in Github that envvar is automatic.

This could probably be replaced by some nice Git-fu instead and some
bash `if` statements. Or maybe
[`tagver`](https://github.com/leosunmo/tagver).