Add support for piping out
This is probably over-engineered, but stdin/stdout is fun.

This adds support for detecting if we're piping out to something as well
as if we're receiving data from stdin.

This means that if we're piping out to something, it won't default to
writing the rendered Dockerfile to the data-file's directory.

This also detects when data is provided to us through stdin, and there's no
output directory provided with `-o` AND there's nothing to send to on
`stdout`. This is considered too ambigious and returns an error.